Paginated S3 Reading
Also cleans up orphan .processed files, and limits number of object
processing routines

+++ main.go
@@ -55,18 +55,37 @@ func main() {
 	}
 	s3Cli := s3.NewFromConfig(cfg)
 
-	// List objects matching bucket / prefix
-	res, err := s3Cli.ListObjects(ctx, &s3.ListObjectsInput{
-		Bucket: aws.String(opts.Bucket),
-		Prefix: aws.String(opts.Prefix),
-	})
-	if err != nil {
-		log.Fatal().Err(err).Msg("Unable to list objects")
+	maxKeys := int32(1000)
+	var elems s3.ListObjectsOutput
+
+	p := s3.NewListObjectsV2Paginator(
+		s3Cli,
+		&s3.ListObjectsV2Input{
+			Bucket:  aws.String(opts.Bucket),
+			Prefix:  aws.String(opts.Prefix),
+			MaxKeys: &maxKeys,
+		},
+		func(o *s3.ListObjectsV2PaginatorOptions) {
+			if v := int32(maxKeys); v != 0 {
+				o.Limit = v
+			}
+		})
+
+	var i int
+	for p.HasMorePages() {
+		i++
+		page, err := p.NextPage(ctx)
+		if err != nil {
+			log.Fatal().Err(err).Msg("Unable to get object page")
+		}
+
+		// Log the objects found
+		elems.Contents = slices.Concat(elems.Contents, page.Contents)
 	}
 
-	unprocCsvs, procCsvs := filterBucketContent(*res, opts.Suffix, opts.ProcessedFlagSuffix)
+	unprocCsvs, procCsvs, orphanFlags := filterBucketContent(elems, opts.Suffix, opts.ProcessedFlagSuffix)
 
-	if len(procCsvs.Contents)+len(unprocCsvs.Contents) == 0 {
+	if len(procCsvs.Contents)+len(unprocCsvs.Contents)+len(orphanFlags.Contents) == 0 {
 		log.Info().Msg("No objects matching bucket / prefix, processing done !")
 		return
 	}
@@ -102,46 +121,70 @@ func main() {
 		})
 	}
 
+	if len(orphanFlags.Contents) > 0 {
+		parallelApply(orphanFlags, func(o types.Object) {
+			if err := cleanObject(ctx, s3Cli, o); err != nil {
+				log.Error().Err(err).Msg("Cleaning error")
+			}
+		})
+	}
+
 	log.Info().
 		Dur("elapsed", time.Since(start)).
 		Msg("Processing ended !")
 }
 
-func filterBucketContent(elems s3.ListObjectsOutput, csvSuffix, processedFlagSuffix string) (unprocessed, processed s3.ListObjectsOutput) {
+func filterBucketContent(elems s3.ListObjectsOutput, csvSuffix, processedFlagSuffix string) (unprocessed, processed, orphanFlags s3.ListObjectsOutput) {
 	// Rely on .processed files present on the bucket to detect which csv
 	// has already been pushed to influx and which has yet to be processed
 	csvFiles := []types.Object{}
-	processedElems := []string{}
+	flags := []types.Object{}
+	objectNames := []string{}
+	flagNames := []string{}
 
 	if len(csvSuffix) == 0 {
-		return unprocessed, processed
+		return unprocessed, processed, orphanFlags
 	}
 	for _, s := range elems.Contents {
 		if strings.HasSuffix(*s.Key, csvSuffix) {
 			csvFiles = append(csvFiles, s)
+			objectNames = append(objectNames, *s.Key)
 		}
 		if strings.HasSuffix(*s.Key, processedFlagSuffix) {
-			processedElems = append(processedElems, strings.ReplaceAll(*s.Key, processedFlagSuffix, csvSuffix))
+			flags = append(flags, s)
+			flagNames = append(flagNames, *s.Key)
 		}
 	}
 
 	for _, o := range csvFiles {
-		if !slices.Contains(processedElems, *o.Key) {
+		flagName := strings.ReplaceAll(*o.Key, csvSuffix, processedFlagSuffix)
+		if !slices.Contains(flagNames, flagName) {
 			unprocessed.Contents = append(unprocessed.Contents, o)
 		} else {
 			processed.Contents = append(processed.Contents, o)
 		}
 	}
 
-	return unprocessed, processed
+	for _, f := range flags {
+		fileName := strings.ReplaceAll(*f.Key, processedFlagSuffix, csvSuffix)
+		if !slices.Contains(objectNames, fileName) {
+			orphanFlags.Contents = append(orphanFlags.Contents, f)
+		}
+
+	}
+
+	return unprocessed, processed, orphanFlags
 }
 
 func parallelApply(list s3.ListObjectsOutput, fn func(o types.Object)) {
+	semaphore := make(chan struct{}, opts.MaxRoutines)
 	var wg sync.WaitGroup
 	wg.Add(len(list.Contents))
 	for _, item := range list.Contents {
 		o := item
 		go func() {
+			semaphore <- struct{}{}        // Lock channel
+			defer func() { <-semaphore }() // Release channel
 			defer wg.Done()
 			fn(o)
 		}()
@@ -238,18 +281,20 @@ func cleanObject(
 			return err
 		}
 
-		markerFileName := strings.ReplaceAll(*o.Key, opts.Suffix, opts.ProcessedFlagSuffix)
-		_, err = s3Cli.DeleteObject(ctx, &s3.DeleteObjectInput{
-			Bucket: aws.String(opts.Bucket),
-			Key:    &markerFileName,
-		})
-		if err != nil {
-			log.Error().
-				Err(err).
-				Str("bucket", opts.Bucket).
-				Str("object", aws.ToString(o.Key)).
-				Msg("Unable to delete object")
-			return err
+		if !strings.Contains(*o.Key, opts.ProcessedFlagSuffix) {
+			markerFileName := strings.ReplaceAll(*o.Key, opts.Suffix, opts.ProcessedFlagSuffix)
+			_, err = s3Cli.DeleteObject(ctx, &s3.DeleteObjectInput{
+				Bucket: aws.String(opts.Bucket),
+				Key:    &markerFileName,
+			})
+			if err != nil {
+				log.Error().
+					Err(err).
+					Str("bucket", opts.Bucket).
+					Str("object", aws.ToString(o.Key)).
+					Msg("Unable to delete object")
+				return err
+			}
 		}
 	}
 	return nil
